Add files via upload
Problem 1 : rotated sorted array: 
Time Complexity: O(log n)
Space Complexity: O(1)
Leet Code: Did not run, will run again and post. But runs  on my g++ C++ compiler. 
Approach:
Found the middle element of the rotated sorted array. Checked if the target is equal to the mid of the array. If yes, returned the index. 
If not, check if the left side of the array, that is from A[0] to A[mid] is sorted. If the target element is found return the index, else check right side of the array.  If the target is not found, return -1. 

Problem 2: Infinite sorted array
Time Complexity: O(log n)
Leet Code: Couldn't find the problem on leet code. 
Space Complexity: O(1)
Approach: Calculate the middle element of the sorted array.  If target == mid, return index.  Let low be the pointer to the 1st element and high be the pointer to the 2nd element. If target is greater than the high element, then copy the high index element to the low index and double the high index. If  the target is lesser than the high index element then apply regular binary search on the low and high indices found.